Cadastrando prodoto, falta o upload da foto

+++ client/produto/produto.js
@@ -7,7 +7,7 @@ Template.ProdutoDetalheLayout.helpers({
 
 Template.ProdutoDetalheLayout.events({
     'click .cart': function (e, template) {
-        var quantidade = parseInt(template.find('input').value);
+        var quantidade = parseFloat(template.find('input').value);
         var produtoId = FlowRouter.getParam('id');
         var userid = Meteor.userId();
         Meteor.call('addAoCarrinho', quantidade, produtoId, userid);
